Allow pcl to be built against system metslib
This patch allows PCL to build against a system installed metslib
where available.  It uses pkgconfig to try to discover metslib.pc,
and when both are available HAVE_METSLIB is set to ON and the
include paths are modified to use metslib's location as discovered
by pkgconfig.  When pkgconfig and/or metslib is not found, cmake
will continue to use the bundled version of metslib as before.