Split PCL_ADD_POINT4D and PCL_ADD_NORMAL4D into parts
PCL_ADD_POINT4D becomes a composite of two simpler macros:

* PCL_ADD_UNION_POINT4D
* PCL_ADD_EIGEN_MAPS_POINT4D

The former adds a union, and the latter adds a family of Eigen vector
accessors. The same holds for PCL_ADD_NORMAL4D. These splits are not
strictly necessary, as both macros are always used together. However,
this way the style consistent with PCL_ADD_..._RGB macros, which we
ought to have separated.