Use flags to check whether member variables have been set
For the person classifier a flag has already been used, but for the
intrinsics matrix and the ground coefficients, data structure specifics
are used in order to determine whether or not the variable has been set.

Since not all data structures have such a specific characteristic,
(e.g. in one of the next patches we introduce a transformation matrix
for which it can't be easily checked whether it has been set or not),
and for consistency, it is better to use flags throughout the whole code.