Add transformations for the point cloud/ground plane/intrinsics matrix
The choice of the coordinate frame matters for the people tracker
(e.g. for the height calculations), and it is important to allow for
different choices of coordinate systems. Doing the transformation inside
the people tracker leads to a better performance than doing it outside
(since it can be applied to the downsampled point cloud), and also
simplifies the use of the people tracker.