build: Fix issue with windows library suffixes
Currently when using PCLConfig.cmake the libraries are searched with a
suffix of either _debug or _release.  When building with either
RelWithDebInfo or MinSizeRel these are not preoprly setup in the
CMakeLists.txt

As result when building with RelWithDebInfo you get libraries named
pcl_io.dll

but the PCLConfig.cmake looks for pcl_io_release.dll

This patch adds the release suffix to both RelWithDebInfo and
MinSizeRel to solve the problem.