Use two-pass mean and covariance calculation to avoid numeric error
The current implementation uses a naive one-pass algorithm
that is subject to significant numeric error.

http://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Na.C3.AFve_algorithm

This occurs especially when the variance is small relative to the mean of the data.
For example, a thin plane far from the origin.

The errors has been observed to impact.

1) Normal calculation filter
2) RANSAC plane segmentation

Possible solutions include a two-pass algorithm or a one-pass algorithm using
a recurrence relation.

The two-pass algorithm was chosen for two reasons

A) Simplicity of implementation and verification
B) Avoid this performance penalty of division within the per-point loop