Test that a translated point cloud has the same normals
If the covariance estimator is using the correct version it should
be robust to a mean shift in the points and provide the same answer.

Note the test of plane position is commented in the translated version
because the full plane equation does shift with a translation, but the
orientation should remain the same.