Add PCLVisualizer::contains().
`contains` allows to check if the viewer (or more precisely the *_actor_map)
already contains input data with the given id.

Why `contains()`:

Currently `updatePointCloud` checks if the viewer contains something with the
given id AND updates the point cloud.  However, in many scenarios only the
check is important. The new `contains` is much clearer and only requires the id
of the data.

`contains` also removes the need for manual bookkeeping (ala `graph_added`
flags),  because now it is easy to query any *_actor_map, not just the
cloud_actor_map.