kinfu LS: shifting: extract data in chunks.
Currently, if the extraction buffer fills up while shifting, the extraction results in incomplete data.
This patch rewrites the extraction procedure so that the kernel saves its state and terminates.
The buffer is then cleared and the kernel resumes from where it left, until all data is extracted.

Removed warp-synchronous programming, which is now unreliable.