Use EXPECT_NEAR to check transformed point coordinates
The Matrix4Affine3Transform case in the `a_transforms_test` fails
consistently in my environment (Eigen 3.2.7, clang 3.7.0) with the
following errors:

3: pcl/test/test_transforms.cpp:342: Failure
3: Value of: ct[0].x
3:   Actual: -1.0601168
3: Expected: pt.x
3: Which is: -1.0601162
3: pcl/test/test_transforms.cpp:351: Failure
3: Value of: ct[0].x
3:   Actual: -1.0601168
3: Expected: pt.x
3: Which is: -1.0601162
3: pcl/test/test_transforms.cpp:356: Failure
3: Value of: ct[0].x
3:   Actual: -1.0601168
3: Expected: pt.x
3: Which is: -1.0601162
3: [  FAILED  ] PCL.Matrix4Affine3Transform (0 ms)

The `EXPECT_FLOAT_EQ` macro is used for checks there. Apparently, a
precision problem. Since most other checks in this test case are done
with `EXPECT_NEAR` and 1e-4 epsilon, I converted the problematic ones
to this style.