testPoint: also return true when adding k-th neighbor
I was perceiving occasional spikes in the computation time for
kNN search. This is due to the search box not being updated when
adding the k-th neighbor and subsequent iterations of the do-while
loop in kNN search never trigger the else case in testPoint. This
way we iterate over way too many points during search.