fix link order issue with boost
This could happen with any boost that was compiled
with -DBOOST_SYSTEM_NO_DEPRECATED.  As of boost-1.66.0
that's the effective default, so the issue is more apparent.
In this case one get an undefined reference to
boost::system::system_category() as the 'system' module
is first in the link order, which will cause that as yet unused
symbol to be discarded by the linker.

Instead move the 'system' module to be the last in the link order
so that all references are resolved.

This was tested with cmake 3.5.2