Refactor SAC models to eliminate tmp_inliers_ member field
This field is used only by OptimizationFunctor inside
optimizeModelCoefficients() function. Since it is logically local for
that function, the pointer to inliers vector is now passed directly to
OptimizationFunction.

Additionally, mark the pointer to self that is passed to
OptimizationFunctor as const, because the functor does not mutate the
state of the model.

After this patch the state of the SAC model is not mutated during the
optimizeModelCoefficients() call and this method may be marked as const.