Fix source file name path stripping in Doxygen
Currently the STRIP_FROM_PATH option is not set, as a result the full
path of source files is "leaked" into the generated HTML. For example,
this documentation page:

http://docs.pointclouds.org/1.9.1/voxel__grid__occlusion__estimation_8h_source.html

Has the following title:

<title>Point Cloud Library (PCL): /__w/1/s/filters/include/pcl/filters/voxel_grid_occlusion_estimation.h Source File</title>

This is an absolute path on Azure Pipelines worker.

We already have path stripping set up properly in STRIP_FROM_INC_PATH,
so this commit just copies the same paths over.

Additionally indentation is fixed in the generated Doxyfile by using
spaces instead of tabs.