Fix compilation of CUDA code on Windows
In pcl_macros.h the attribute [[deprecated]] is used. It seems that
nvcc doesn't support this attribute under Windows at the moment.
I tried it with CUDA Toolkit 10.1 with Windows 10. I also checked the
release notes of 10.2 but they didn't address it there.

To fix this I introduced a macro PCL_DEPRECATED that handles the details
for the different platforms and compilers.