Allow SAC segmentation to take a pre-seeded RNG instead of always seeding to 12345u or current time
Existing SAC classes take a boolean "random" parameter in the constructor, and
always create a new RNG and seeds it to 12345u or the current time. This means
that all SAC operations that take place within the same second will yield the
same result. This prevents performing multiple fast segmentation operations,
and prevents performing multiple segmentations in parallel.

This change adds a new "rng_gen" parameter to all SAC model constructors so
that the caller can pass in a pre-seeded RNG, or re-use the RNG across
multiple operations.

This change should be backwards compatible because all parameters and template
parameters have sensible defaults. The tests pass with no changes required.
I created a separate BoostUniformGenerator which uses the boost random classes,
so that they would generate the values expected by the tests, and prove that
this change remains correct. In the future, we can change to using
UniformGenerator and update the test expectations.