Fix problem with spin() and spinOnce() for X Window System (#5252)
* Fix problem with spin() and spinOnce() for X Window System
Due to a problem in VTK (versions 9.0.2, 9.0.3, and 9.1.0), the Start() function of the vtkXRenderWindowInteractor crashes when Terminate() is called. This leads to spin() and spinOnce() crashing.
This commit adds a fixed version of the vtkXRenderWindowInteractor that is used for the previously mentioned VTK versions.
Additionally, it is necessary to change spinOnce() from Start()+timer+Terminate() to ProcessEvents()+sleep since Terminate() closes the window in future VTK versions (and the fixed vtkXRenderWindowInteractor).
The changes are limited to the affected versions and the X Window System to avoid regressions for e.g. Windows and MacOS

* Do not always compile the fixed X rwi

* Only include header if necessary