Reimplement boost::split and optimize tokenization
This commit reimplements boost::split in a io/split.h. Implementation is
then used for obj and pcd file parsing. Boost::split was reimplemented
because the function was aggressively allocating new memory in a hot
parsing loops and that was slowing down sequential and parallel
execution.

New implementation is generic in way that it will work with current
std::vector<std::string> result type but also with a
std::vector<std::string_view> result type.

Signed-off-by: Darko Janekovic <darko.janekovic@gmail.com>