Point Cloud Editor: fix select2D after switch to QOpenGLWidget (#5685)
* Point Cloud Editor: fix select2D after switch to QOpenGLWidget
In commit https://github.com/PointCloudLibrary/pcl/commit/38cbd621d0632895e1bed04841a612f97f574abc the CloudEditorWidget was switched from QGLWidget to QOpenGLWidget to be compatible with Qt6.
The select2D tool needs the viewport and the projection matrix, which are loaded with `glGetIntegerv(GL_VIEWPORT,viewport)` and `glGetFloatv(GL_PROJECTION_MATRIX, project)` (they are set in `CloudEditorWidget::resizeGL`). However, since the switch to QOpenGLWidget, the received projection matrix is just the identity matrix, and the viewport does not look correct either. It seems that QOpenGLWidget overwrites them.
These changes introduce variables in CloudEditorWidget to safely store the viewport and projection matrix, and passes a function to the select2D tool to ask for them when needed.

* Add include for windows

* Additional fix for hdpi mac displays