Make OpenMP available in downstream projects (#5744)
* Make OpenMP available in downstream projects
Previously, if a downstream project used PCL_NO_PRECOMPILE or if it used a PCL class which includes OpenMP code, but PCL does not make an instantiation of that class, OpenMP was not available.
In other words, when PCL code using OpenMP was compiled when building PCL, OpenMP was available. However when PCL code using OpenMP was compiled in the downstream/user project (e.g. in the two mentioned cases), OpenMP was not available.
The fix is to search for OpenMP again in PCLConfig.cmake

* OpenMP as dep for more modules