Use real ALIAS(cmake >=3.18) target for flann to preserve properties on original target. (#5861)
* Use real ALIAS target for flann to preserve properties on original target.

With the current method all properties set on the flann::flann_cpp and flann::flann_cpp target are lost, as they are not copied to the new FLANN::FLANN interface target. However, there are cases, where we would like preserve these, e.g. to keep the INTERFACE_INCLUDE_DIRECTORIES property that includes the path to the LZ4 headers. Therefore this patch first checks for the appropriate SHARED or STATIC version and then aliases FLANN::FLANN to the respective library.

* Use appropriate comparison operator

Co-authored-by: Lars Glud <larshg@gmail.com>

* Only use ALIAS Target for CMAKE >= 3.18

* Use appropriate comparison operator

Co-authored-by: Markus Vieth <39675748+mvieth@users.noreply.github.com>

---------

Co-authored-by: Lars Glud <larshg@gmail.com>
Co-authored-by: Markus Vieth <39675748+mvieth@users.noreply.github.com>